+++ src/compiler/app-core/app-data.ts
@@ -32,6 +32,7 @@ export const getBuildFeatures = (cmps: ComponentCompilerMeta[]): BuildFeatures =
     cmpWillLoad: cmps.some((c) => c.hasComponentWillLoadFn),
     cmpWillUpdate: cmps.some((c) => c.hasComponentWillUpdateFn),
     cmpWillRender: cmps.some((c) => c.hasComponentWillRenderFn),
+    formAssociated: cmps.some((c) => c.formAssociated),
 
     connectedCallback: cmps.some((c) => c.hasConnectedCallbackFn),
     disconnectedCallback: cmps.some((c) => c.hasDisconnectedCallbackFn),
@@ -167,7 +168,7 @@ export const updateBuildConditionals = (config: ValidatedConfig, b: BuildConditi
   // TODO(STENCIL-914): remove this option when `experimentalSlotFixes` is the default behavior
   b.slotChildNodesFix = config.extras.slotChildNodesFix;
   // TODO(STENCIL-914): remove this option when `experimentalSlotFixes` is the default behavior
-  b.patchPseudoShadowDom = config.extras.experimentalSlotFixes;
+  b.experimentalSlotFixes = config.extras.experimentalSlotFixes;
   // TODO(STENCIL-914): remove this option when `experimentalSlotFixes` is the default behavior
   b.cloneNodeFix = config.extras.cloneNodeFix;
   b.lifecycleDOMEvents = !!(b.isDebug || config._isTesting || config.extras.lifecycleDOMEvents);