+++ src/utils/path.ts
@@ -165,7 +165,8 @@ export const normalizeFsPath = (p: string) => normalizePath(p.split('?')[0].repl
 export const normalizeFsPathQuery = (importPath: string) => {
   const pathParts = importPath.split('?');
   const filePath = normalizePath(pathParts[0]);
-  const ext = filePath.split('.').pop().toLowerCase();
+  const filePathParts = filePath.split('.');
+  const ext = filePathParts.length > 1 ? filePathParts.pop()!.toLowerCase() : null;
   const params = pathParts.length > 1 ? new URLSearchParams(pathParts[1]) : null;
   const format = params ? params.get('format') : null;
 
@@ -216,4 +217,32 @@ export function relative(from: string, to: string): string {
  */
 export function join(...paths: string[]): string {
   return normalizePath(path.join(...paths), false);
+}
+
+/**
+ * A wrapped version of node.js' {@link path.resolve} which adds our custom
+ * normalization logic. This resolves a path to a given (relative or absolute)
+ * path.
+ *
+ * @throws the underlying node function will throw if any argument is not a
+ * string
+ * @param paths a path or path fragments to resolve
+ * @returns a resolved path!
+ */
+export function resolve(...paths: string[]): string {
+  return normalizePath(path.resolve(...paths), false);
+}
+
+/**
+ * A wrapped version of node.js' {@link path.normalize} which adds our custom
+ * normalization logic. This normalizes a path, de-duping repeated segment
+ * separators and resolving `'..'` segments.
+ *
+ * @throws the underlying node function will throw if the argument is not a
+ * string
+ * @param toNormalize a path to normalize
+ * @returns a normalized path!
+ */
+export function normalize(toNormalize: string): string {
+  return normalizePath(path.normalize(toNormalize), false);
 }